Changed: Split up the functionality into the multiple files, each with a set of tests.
